+++ Client/src/components/Brews/tempdb.js
@@ -1,0 +1,20 @@
+const brews = [
+    ///provisional use database instead//fetch from  db
+    {
+        id: 1,
+        name: "french press",
+        url: "/french-press",
+    },
+    {
+        id: 2,
+        name: "italian press",
+        url: "/italian-press",
+    },
+    {
+        id: 3,
+        name: "german press",
+        url: "/german-press",
+    }
+]
+
+export default brews;
